rename bounding_box_wrapper.py
